requirements.txt
Flask>=2.3.2,<3.0
psutil>=5.9.5
redis>=4.5.5
pandas>=2.0.0
torch>=2.5.0,<3.0
websockets>=11.0.3
requests>=2.31.0
PyYAML>=6.0
Flask-SocketIO==5.3.6
eventlet>=0.33.0
numpy>=1.26.4
pytest>=7.0.0
pytest-cov>=4.0.0
coverage>=7.0.0